feat: Add functionality to request data by quantity specified
LTA DataMall API returns up to 500 values (some endpoints have different behaviour) at a time. This means multiple requests are needed for more data, which is needs the $skip parameter in the url.

This feature streamlines the process by including the logic for that in the code, simplifying calls to the API Client